add ui widget to switch between csaf 2.0 and 2.1
