+++ app/lib/app/SecvisogramPage/View/FormEditorTab.js
@@ -1,11 +1,17 @@
 import React from 'react'
+import { uiSchemas } from '../../../uiSchemas.js'
 import { GenericEditor } from './FormEditor/editors.js'
-import schema from './FormEditor/schema.js'
 
-export default function FormEditorTab() {
-  const property = /** @type {import('./FormEditor/schema.js').Property} */ (
-    schema
-  )
+/**
+ * @typedef {object} Props
+ * @property {import('#lib/uiSchemas.js').UiSchemaVersion} schemaVersion
+ */
+
+/**
+ * @type {React.FC<Props>}
+ */
+export const FormEditorTab = ({ schemaVersion }) => {
+  const property = uiSchemas[schemaVersion].content
 
   return (
     <div className="flex h-full w-full grow">
