+++ app/lib/uiSchemas/csaf_2_0.js
@@ -1,0 +1,2 @@
+export { default as content } from './csaf_2_0/content.js'
+export { default as metaData } from './csaf_2_0/metaData.js'
