+++ app/lib/uiSchemas/csaf_2_0/content.js
@@ -1,4 +1,4 @@
-/** @typedef {import('./shared/types').Property} Property */
+/** @type {import('#lib/app/SecvisogramPage/shared/types.js').Property} */
 export default /** @type {const} */ ({
   key: '',
   fullName: [],
@@ -10,33 +10,6 @@ export default /** @type {const} */ ({
   type: 'OBJECT',
   metaInfo: {
     propertyList: [
-      {
-        key: '$schema',
-        fullName: ['$schema'],
-        title: 'JSON schema',
-        description:
-          'Contains the URL of the CSAF JSON schema which the document promises to be valid for.',
-        metaData: {
-          uiType: 'STRING_URI',
-          userDocumentation: {
-            specification: 'docs/user/document-spec.en.md',
-            usage: 'docs/user/document-usage.en.md',
-          },
-          i18n: { title: 'SchemaTitle', description: 'SchemaDescription' },
-          relevanceLevels: {
-            csaf_base: 'mandatory',
-            csaf_security_incident_response: 'mandatory',
-            csaf_informational_advisory: 'mandatory',
-            csaf_security_advisory: 'mandatory',
-            csaf_vex: 'mandatory',
-          },
-        },
-        enum: [
-          'https://docs.oasis-open.org/csaf/csaf/v2.1/csaf_json_schema.json',
-        ],
-        metaInfo: {},
-        type: 'STRING',
-      },
       {
         key: 'document',
         fullName: ['document'],
@@ -475,7 +448,7 @@ export default /** @type {const} */ ({
                   description: 'DocumentCsafVersionDescription',
                 },
               },
-              enum: ['2.1'],
+              enum: ['2.0'],
               metaInfo: {},
               type: 'STRING',
             },
@@ -505,119 +478,6 @@ export default /** @type {const} */ ({
               type: 'OBJECT',
               metaInfo: {
                 propertyList: [
-                  {
-                    key: 'sharing_group',
-                    fullName: ['document', 'distribution', 'sharing_group'],
-                    title: 'Sharing Group',
-                    description:
-                      'Contains information about the group this document is intended to be shared with.',
-                    metaData: {
-                      userDocumentation: {
-                        specification:
-                          'docs/user/document/distribution/sharing_group-spec.en.md',
-                        usage:
-                          'docs/user/document/distribution/sharing_group-usage.en.md',
-                      },
-                      relevanceLevels: {
-                        csaf_base: 'want_to_have',
-                        csaf_security_incident_response: 'want_to_have',
-                        csaf_informational_advisory: 'want_to_have',
-                        csaf_security_advisory: 'best_practice',
-                        csaf_vex: 'want_to_have',
-                      },
-                      i18n: {
-                        title: 'DocumentDistributionSharingGroupTitle',
-                        description:
-                          'DocumentDistributionSharingGroupDescription',
-                      },
-                    },
-                    type: 'OBJECT',
-                    metaInfo: {
-                      propertyList: [
-                        {
-                          key: 'id',
-                          fullName: [
-                            'document',
-                            'distribution',
-                            'sharing_group',
-                            'id',
-                          ],
-                          title: 'Sharing Group ID',
-                          description:
-                            'Provides the unique ID for the sharing group.',
-                          metaData: {
-                            uiType: 'STRING_ENUM',
-                            userDocumentation: {
-                              specification:
-                                'docs/user/document/distribution/sharing_group/id-spec.en.md',
-                              usage:
-                                'docs/user/document/distribution/sharing_group/id-usage.en.md',
-                            },
-                            relevanceLevels: {
-                              csaf_base: 'mandatory',
-                              csaf_security_incident_response: 'mandatory',
-                              csaf_informational_advisory: 'mandatory',
-                              csaf_security_advisory: 'mandatory',
-                              csaf_vex: 'mandatory',
-                            },
-                            i18n: {
-                              title: 'DocumentDistributionSharingGroupIdTitle',
-                              description:
-                                'DocumentDistributionSharingGroupIdDescription',
-                            },
-                          },
-                          pattern:
-                            '^(([0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[0-9a-f]{4}-[0-9a-f]{12})|([0]{8}-([0]{4}-){3}[0]{12})|([f]{8}-([f]{4}-){3}[f]{12}))$',
-                          metaInfo: {},
-                          type: 'STRING',
-                        },
-                        {
-                          key: 'name',
-                          fullName: [
-                            'document',
-                            'distribution',
-                            'sharing_group',
-                            'name',
-                          ],
-                          title: 'Sharing Group Name',
-                          description:
-                            'Contains a human-readable name for the sharing group.',
-                          metaData: {
-                            userDocumentation: {
-                              specification:
-                                'docs/user/document/distribution/sharing_group/name-spec.en.md',
-                              usage:
-                                'docs/user/document/distribution/sharing_group/name-usage.en.md',
-                            },
-                            relevanceLevels: {
-                              csaf_base: 'nice_to_know',
-                              csaf_security_incident_response: 'nice_to_know',
-                              csaf_informational_advisory: 'nice_to_know',
-                              csaf_security_advisory: 'want_to_have',
-                              csaf_vex: 'nice_to_know',
-                            },
-                            i18n: {
-                              title:
-                                'DocumentDistributionSharingGroupNameTitle',
-                              description:
-                                'DocumentDistributionSharingGroupNameDescription',
-                            },
-                          },
-                          minLength: 1,
-                          examples: [
-                            'Customer A',
-                            'ISAC members',
-                            'NIS2 regulated important entities in Germany, sector water',
-                            'Pre-Sharing group for advisory discussion',
-                            'Users of Product A',
-                            'US Federal Civilian Authorities',
-                          ],
-                          metaInfo: {},
-                          type: 'STRING',
-                        },
-                      ],
-                    },
-                  },
                   {
                     key: 'text',
                     fullName: ['document', 'distribution', 'text'],
@@ -713,14 +573,7 @@ export default /** @type {const} */ ({
                                 'DocumentDistributionTlpLabelDescription',
                             },
                           },
-                          default: 'CLEAR',
-                          enum: [
-                            'AMBER',
-                            'AMBER+STRICT',
-                            'CLEAR',
-                            'GREEN',
-                            'RED',
-                          ],
+                          enum: ['AMBER', 'GREEN', 'RED', 'WHITE'],
                           metaInfo: {},
                           type: 'STRING',
                         },
@@ -1045,7 +898,6 @@ export default /** @type {const} */ ({
                     enum: [
                       'coordinator',
                       'discoverer',
-                      'multiplier',
                       'other',
                       'translator',
                       'user',
@@ -1744,7 +1596,7 @@ export default /** @type {const} */ ({
                     fullName: ['document', 'tracking', 'initial_release_date'],
                     title: 'Initial release date',
                     description:
-                      'The date when this document was first released to the specified target group.',
+                      'The date when this document was first published.',
                     metaData: {
                       uiType: 'STRING_DATETIME',
                       userDocumentation: {
@@ -2205,7 +2057,6 @@ export default /** @type {const} */ ({
                           'language',
                           'legacy',
                           'patch_level',
-                          'platform',
                           'product_family',
                           'product_name',
                           'product_version',
@@ -2428,7 +2279,7 @@ export default /** @type {const} */ ({
                                     },
                                     minLength: 5,
                                     pattern:
-                                      '^((cpe:2\\.3:[aho\\*\\-](:(((\\?*|\\*?)([a-zA-Z0-9\\-\\._]|(\\\\[\\\\\\*\\?!"#\\$%&\'\\(\\)\\+,\\/:;<=>@\\[\\]\\^`\\{\\|\\}~]))+(\\?*|\\*?))|[\\*\\-])){5}(:(([a-zA-Z]{2,3}(-([a-zA-Z]{2}|[0-9]{3}))?)|[\\*\\-]))(:(((\\?*|\\*?)([a-zA-Z0-9\\-\\._]|(\\\\[\\\\\\*\\?!"#\\$%&\'\\(\\)\\+,\\/:;<=>@\\[\\]\\^`\\{\\|\\}~]))+(\\?*|\\*?))|[\\*\\-])){4})|([c][pP][eE]:\\/[AHOaho]?(:[A-Za-z0-9\\._\\-~%]*){0,6}))$',
+                                      '^(cpe:2\\.3:[aho\\*\\-](:(((\\?*|\\*?)([a-zA-Z0-9\\-\\._]|(\\\\[\\\\\\*\\?!"#\\$%&\'\\(\\)\\+,/:;<=>@\\[\\]\\^`\\{\\|\\}~]))+(\\?*|\\*?))|[\\*\\-])){5}(:(([a-zA-Z]{2,3}(-([a-zA-Z]{2}|[0-9]{3}))?)|[\\*\\-]))(:(((\\?*|\\*?)([a-zA-Z0-9\\-\\._]|(\\\\[\\\\\\*\\?!"#\\$%&\'\\(\\)\\+,/:;<=>@\\[\\]\\^`\\{\\|\\}~]))+(\\?*|\\*?))|[\\*\\-])){4})|([c][pP][eE]:/[AHOaho]?(:[A-Za-z0-9\\._\\-~%]*){0,6})$',
                                     metaInfo: {},
                                     type: 'STRING',
                                   },
@@ -2840,23 +2691,24 @@ export default /** @type {const} */ ({
                                     },
                                   },
                                   {
-                                    key: 'purls',
+                                    key: 'purl',
                                     fullName: [
                                       'product_tree',
                                       'branches',
                                       'product',
                                       'product_identification_helper',
-                                      'purls',
+                                      'purl',
                                     ],
-                                    title: 'List of package URLs',
+                                    title: 'package URL representation',
                                     description:
-                                      'Contains a list of package URLs (purl).',
+                                      'The package URL (purl) attribute refers to a method for reliably identifying and locating software packages external to this specification.',
                                     metaData: {
+                                      uiType: 'STRING_URI',
                                       userDocumentation: {
                                         specification:
-                                          'docs/user/types/full_product_name/product_identification_helper/purls-spec.en.md',
+                                          'docs/user/types/full_product_name/product_identification_helper/purl-spec.en.md',
                                         usage:
-                                          'docs/user/types/full_product_name/product_identification_helper/purls-usage.en.md',
+                                          'docs/user/types/full_product_name/product_identification_helper/purl-usage.en.md',
                                       },
                                       relevanceLevels: {
                                         csaf_base: 'nice_to_know',
@@ -2874,52 +2726,11 @@ export default /** @type {const} */ ({
                                           'FullProductNameProductIdentificationHelperPurlDescription',
                                       },
                                     },
-                                    type: 'ARRAY',
-                                    metaInfo: {
-                                      arrayType: {
-                                        key: '',
-                                        fullName: [
-                                          'product_tree',
-                                          'branches',
-                                          'product',
-                                          'product_identification_helper',
-                                          'purls',
-                                        ],
-                                        title: 'package URL representation',
-                                        description:
-                                          'The package URL (purl) attribute refers to a method for reliably identifying and locating software packages external to this specification.',
-                                        metaData: {
-                                          uiType: 'STRING_URI',
-                                          userDocumentation: {
-                                            specification:
-                                              'docs/user/types/full_product_name/product_identification_helper/purls-spec.en.md',
-                                            usage:
-                                              'docs/user/types/full_product_name/product_identification_helper/purls-usage.en.md',
-                                          },
-                                          relevanceLevels: {
-                                            csaf_base: 'nice_to_know',
-                                            csaf_security_incident_response:
-                                              'nice_to_know',
-                                            csaf_informational_advisory:
-                                              'nice_to_know',
-                                            csaf_security_advisory:
-                                              'nice_to_know',
-                                            csaf_vex: 'nice_to_know',
-                                          },
-                                          i18n: {
-                                            title:
-                                              'FullProductNameProductIdentificationHelperPurlsTitle',
-                                            description:
-                                              'FullProductNameProductIdentificationHelperPurlsDescription',
-                                          },
-                                        },
-                                        minLength: 7,
-                                        pattern:
-                                          '^pkg:[A-Za-z\\.\\-\\+][A-Za-z0-9\\.\\-\\+]*\\/.+',
-                                        metaInfo: {},
-                                        type: 'STRING',
-                                      },
-                                    },
+                                    minLength: 7,
+                                    pattern:
+                                      '^pkg:[A-Za-z\\.\\-\\+][A-Za-z0-9\\.\\-\\+]*/.+',
+                                    metaInfo: {},
+                                    type: 'STRING',
                                   },
                                   {
                                     key: 'sbom_urls',
@@ -3549,7 +3360,7 @@ export default /** @type {const} */ ({
                               },
                               minLength: 5,
                               pattern:
-                                '^((cpe:2\\.3:[aho\\*\\-](:(((\\?*|\\*?)([a-zA-Z0-9\\-\\._]|(\\\\[\\\\\\*\\?!"#\\$%&\'\\(\\)\\+,\\/:;<=>@\\[\\]\\^`\\{\\|\\}~]))+(\\?*|\\*?))|[\\*\\-])){5}(:(([a-zA-Z]{2,3}(-([a-zA-Z]{2}|[0-9]{3}))?)|[\\*\\-]))(:(((\\?*|\\*?)([a-zA-Z0-9\\-\\._]|(\\\\[\\\\\\*\\?!"#\\$%&\'\\(\\)\\+,\\/:;<=>@\\[\\]\\^`\\{\\|\\}~]))+(\\?*|\\*?))|[\\*\\-])){4})|([c][pP][eE]:\\/[AHOaho]?(:[A-Za-z0-9\\._\\-~%]*){0,6}))$',
+                                '^(cpe:2\\.3:[aho\\*\\-](:(((\\?*|\\*?)([a-zA-Z0-9\\-\\._]|(\\\\[\\\\\\*\\?!"#\\$%&\'\\(\\)\\+,/:;<=>@\\[\\]\\^`\\{\\|\\}~]))+(\\?*|\\*?))|[\\*\\-])){5}(:(([a-zA-Z]{2,3}(-([a-zA-Z]{2}|[0-9]{3}))?)|[\\*\\-]))(:(((\\?*|\\*?)([a-zA-Z0-9\\-\\._]|(\\\\[\\\\\\*\\?!"#\\$%&\'\\(\\)\\+,/:;<=>@\\[\\]\\^`\\{\\|\\}~]))+(\\?*|\\*?))|[\\*\\-])){4})|([c][pP][eE]:/[AHOaho]?(:[A-Za-z0-9\\._\\-~%]*){0,6})$',
                               metaInfo: {},
                               type: 'STRING',
                             },
@@ -3941,22 +3752,23 @@ export default /** @type {const} */ ({
                               },
                             },
                             {
-                              key: 'purls',
+                              key: 'purl',
                               fullName: [
                                 'product_tree',
                                 'full_product_names',
                                 'product_identification_helper',
-                                'purls',
+                                'purl',
                               ],
-                              title: 'List of package URLs',
+                              title: 'package URL representation',
                               description:
-                                'Contains a list of package URLs (purl).',
+                                'The package URL (purl) attribute refers to a method for reliably identifying and locating software packages external to this specification.',
                               metaData: {
+                                uiType: 'STRING_URI',
                                 userDocumentation: {
                                   specification:
-                                    'docs/user/types/full_product_name/product_identification_helper/purls-spec.en.md',
+                                    'docs/user/types/full_product_name/product_identification_helper/purl-spec.en.md',
                                   usage:
-                                    'docs/user/types/full_product_name/product_identification_helper/purls-usage.en.md',
+                                    'docs/user/types/full_product_name/product_identification_helper/purl-usage.en.md',
                                 },
                                 relevanceLevels: {
                                   csaf_base: 'nice_to_know',
@@ -3973,50 +3785,11 @@ export default /** @type {const} */ ({
                                     'FullProductNameProductIdentificationHelperPurlDescription',
                                 },
                               },
-                              type: 'ARRAY',
-                              metaInfo: {
-                                arrayType: {
-                                  key: '',
-                                  fullName: [
-                                    'product_tree',
-                                    'full_product_names',
-                                    'product_identification_helper',
-                                    'purls',
-                                  ],
-                                  title: 'package URL representation',
-                                  description:
-                                    'The package URL (purl) attribute refers to a method for reliably identifying and locating software packages external to this specification.',
-                                  metaData: {
-                                    uiType: 'STRING_URI',
-                                    userDocumentation: {
-                                      specification:
-                                        'docs/user/types/full_product_name/product_identification_helper/purls-spec.en.md',
-                                      usage:
-                                        'docs/user/types/full_product_name/product_identification_helper/purls-usage.en.md',
-                                    },
-                                    relevanceLevels: {
-                                      csaf_base: 'nice_to_know',
-                                      csaf_security_incident_response:
-                                        'nice_to_know',
-                                      csaf_informational_advisory:
-                                        'nice_to_know',
-                                      csaf_security_advisory: 'nice_to_know',
-                                      csaf_vex: 'nice_to_know',
-                                    },
-                                    i18n: {
-                                      title:
-                                        'FullProductNameProductIdentificationHelperPurlsTitle',
-                                      description:
-                                        'FullProductNameProductIdentificationHelperPurlsDescription',
-                                    },
-                                  },
-                                  minLength: 7,
-                                  pattern:
-                                    '^pkg:[A-Za-z\\.\\-\\+][A-Za-z0-9\\.\\-\\+]*\\/.+',
-                                  metaInfo: {},
-                                  type: 'STRING',
-                                },
-                              },
+                              minLength: 7,
+                              pattern:
+                                '^pkg:[A-Za-z\\.\\-\\+][A-Za-z0-9\\.\\-\\+]*/.+',
+                              metaInfo: {},
+                              type: 'STRING',
                             },
                             {
                               key: 'sbom_urls',
@@ -4909,7 +4682,7 @@ export default /** @type {const} */ ({
                                     },
                                     minLength: 5,
                                     pattern:
-                                      '^((cpe:2\\.3:[aho\\*\\-](:(((\\?*|\\*?)([a-zA-Z0-9\\-\\._]|(\\\\[\\\\\\*\\?!"#\\$%&\'\\(\\)\\+,\\/:;<=>@\\[\\]\\^`\\{\\|\\}~]))+(\\?*|\\*?))|[\\*\\-])){5}(:(([a-zA-Z]{2,3}(-([a-zA-Z]{2}|[0-9]{3}))?)|[\\*\\-]))(:(((\\?*|\\*?)([a-zA-Z0-9\\-\\._]|(\\\\[\\\\\\*\\?!"#\\$%&\'\\(\\)\\+,\\/:;<=>@\\[\\]\\^`\\{\\|\\}~]))+(\\?*|\\*?))|[\\*\\-])){4})|([c][pP][eE]:\\/[AHOaho]?(:[A-Za-z0-9\\._\\-~%]*){0,6}))$',
+                                      '^(cpe:2\\.3:[aho\\*\\-](:(((\\?*|\\*?)([a-zA-Z0-9\\-\\._]|(\\\\[\\\\\\*\\?!"#\\$%&\'\\(\\)\\+,/:;<=>@\\[\\]\\^`\\{\\|\\}~]))+(\\?*|\\*?))|[\\*\\-])){5}(:(([a-zA-Z]{2,3}(-([a-zA-Z]{2}|[0-9]{3}))?)|[\\*\\-]))(:(((\\?*|\\*?)([a-zA-Z0-9\\-\\._]|(\\\\[\\\\\\*\\?!"#\\$%&\'\\(\\)\\+,/:;<=>@\\[\\]\\^`\\{\\|\\}~]))+(\\?*|\\*?))|[\\*\\-])){4})|([c][pP][eE]:/[AHOaho]?(:[A-Za-z0-9\\._\\-~%]*){0,6})$',
                                     metaInfo: {},
                                     type: 'STRING',
                                   },
@@ -5321,18 +5094,19 @@ export default /** @type {const} */ ({
                                     },
                                   },
                                   {
-                                    key: 'purls',
+                                    key: 'purl',
                                     fullName: [
                                       'product_tree',
                                       'relationships',
                                       'full_product_name',
                                       'product_identification_helper',
-                                      'purls',
+                                      'purl',
                                     ],
-                                    title: 'List of package URLs',
+                                    title: 'package URL representation',
                                     description:
-                                      'Contains a list of package URLs (purl).',
+                                      'The package URL (purl) attribute refers to a method for reliably identifying and locating software packages external to this specification.',
                                     metaData: {
+                                      uiType: 'STRING_URI',
                                       userDocumentation: {
                                         specification:
                                           'docs/user/types/full_product_name/product_identification_helper/purl-spec.en.md',
@@ -5355,52 +5129,11 @@ export default /** @type {const} */ ({
                                           'FullProductNameProductIdentificationHelperPurlDescription',
                                       },
                                     },
-                                    type: 'ARRAY',
-                                    metaInfo: {
-                                      arrayType: {
-                                        key: '',
-                                        fullName: [
-                                          'product_tree',
-                                          'relationships',
-                                          'full_product_name',
-                                          'product_identification_helper',
-                                          'purls',
-                                        ],
-                                        title: 'package URL representation',
-                                        description:
-                                          'The package URL (purl) attribute refers to a method for reliably identifying and locating software packages external to this specification.',
-                                        metaData: {
-                                          uiType: 'STRING_URI',
-                                          userDocumentation: {
-                                            specification:
-                                              'docs/user/types/full_product_name/product_identification_helper/purl-spec.en.md',
-                                            usage:
-                                              'docs/user/types/full_product_name/product_identification_helper/purl-usage.en.md',
-                                          },
-                                          relevanceLevels: {
-                                            csaf_base: 'nice_to_know',
-                                            csaf_security_incident_response:
-                                              'nice_to_know',
-                                            csaf_informational_advisory:
-                                              'nice_to_know',
-                                            csaf_security_advisory:
-                                              'nice_to_know',
-                                            csaf_vex: 'nice_to_know',
-                                          },
-                                          i18n: {
-                                            title:
-                                              'FullProductNameProductIdentificationHelperPurlsTitle',
-                                            description:
-                                              'FullProductNameProductIdentificationHelperPurlsDescription',
-                                          },
-                                        },
-                                        minLength: 7,
-                                        pattern:
-                                          '^pkg:[A-Za-z\\.\\-\\+][A-Za-z0-9\\.\\-\\+]*\\/.+',
-                                        metaInfo: {},
-                                        type: 'STRING',
-                                      },
-                                    },
+                                    minLength: 7,
+                                    pattern:
+                                      '^pkg:[A-Za-z\\.\\-\\+][A-Za-z0-9\\.\\-\\+]*/.+',
+                                    metaInfo: {},
+                                    type: 'STRING',
                                   },
                                   {
                                     key: 'sbom_urls',
@@ -6280,16 +6013,17 @@ export default /** @type {const} */ ({
                   type: 'STRING',
                 },
                 {
-                  key: 'cwes',
-                  fullName: ['vulnerabilities', 'cwes'],
-                  title: 'List of CWEs',
-                  description: 'Contains a list of CWEs.',
+                  key: 'cwe',
+                  fullName: ['vulnerabilities', 'cwe'],
+                  title: 'CWE',
+                  description:
+                    'Holds the MITRE standard Common Weakness Enumeration (CWE) for the weakness associated.',
                   metaData: {
                     userDocumentation: {
                       specification:
-                        'docs/user/vulnerabilities/vulnerability/cwes-spec.en.md',
+                        'docs/user/vulnerabilities/vulnerability/cwe-spec.en.md',
                       usage:
-                        'docs/user/vulnerabilities/vulnerability/cwes-usage.en.md',
+                        'docs/user/vulnerabilities/vulnerability/cwe-usage.en.md',
                     },
                     relevanceLevels: {
                       csaf_base: 'want_to_have',
@@ -6299,164 +6033,81 @@ export default /** @type {const} */ ({
                       csaf_vex: 'want_to_have',
                     },
                     i18n: {
-                      title: 'VulnerabilitiesItemsCwesTitle',
-                      description: 'VulnerabilitiesItemsCwesDescription',
+                      title: 'VulnerabilitiesItemsCweTitle',
+                      description: 'VulnerabilitiesItemsCweDescription',
                     },
                     uiType: 'OBJECT_CWE',
                   },
-                  type: 'ARRAY',
+                  type: 'OBJECT',
                   metaInfo: {
-                    arrayType: {
-                      key: '',
-                      fullName: ['vulnerabilities', 'cwes'],
-                      title: 'CWE',
-                      description:
-                        'Holds the MITRE standard Common Weakness Enumeration (CWE) for the weakness associated.',
-                      metaData: {
-                        relevanceLevels: {
-                          csaf_base: 'want_to_have',
-                          csaf_security_incident_response: 'optional',
-                          csaf_informational_advisory: 'excluded',
-                          csaf_security_advisory: 'best_practice',
-                          csaf_vex: 'want_to_have',
+                    propertyList: [
+                      {
+                        key: 'id',
+                        fullName: ['vulnerabilities', 'cwe', 'id'],
+                        title: 'Weakness ID',
+                        description:
+                          'Holds the ID for the weakness associated.',
+                        metaData: {
+                          userDocumentation: {
+                            specification:
+                              'docs/user/vulnerabilities/vulnerability/cwe/id-spec.en.md',
+                            usage:
+                              'docs/user/vulnerabilities/vulnerability/cwe/id-usage.en.md',
+                          },
+                          relevanceLevels: {
+                            csaf_base: 'mandatory',
+                            csaf_security_incident_response: 'mandatory',
+                            csaf_informational_advisory: 'mandatory',
+                            csaf_security_advisory: 'mandatory',
+                            csaf_vex: 'mandatory',
+                          },
+                          i18n: {
+                            title: 'VulnerabilitiesItemsCweIdTitle',
+                            description: 'VulnerabilitiesItemsCweIdDescription',
+                          },
                         },
+                        examples: ['CWE-22', 'CWE-352', 'CWE-79'],
+                        pattern: '^CWE-[1-9]\\d{0,5}$',
+                        metaInfo: {},
+                        type: 'STRING',
                       },
-                      type: 'OBJECT',
-                      metaInfo: {
-                        propertyList: [
-                          {
-                            key: 'id',
-                            fullName: ['vulnerabilities', 'cwes', 'id'],
-                            title: 'Weakness ID',
-                            description:
-                              'Holds the ID for the weakness associated.',
-                            metaData: {
-                              userDocumentation: {
-                                specification:
-                                  'docs/user/vulnerabilities/vulnerability/cwes/id-spec.en.md',
-                                usage:
-                                  'docs/user/vulnerabilities/vulnerability/cwes/id-usage.en.md',
-                              },
-                              relevanceLevels: {
-                                csaf_base: 'mandatory',
-                                csaf_security_incident_response: 'mandatory',
-                                csaf_informational_advisory: 'mandatory',
-                                csaf_security_advisory: 'mandatory',
-                                csaf_vex: 'mandatory',
-                              },
-                              i18n: {
-                                title: 'VulnerabilitiesItemsCwesIdTitle',
-                                description:
-                                  'VulnerabilitiesItemsCwesIdDescription',
-                              },
-                            },
-                            examples: ['CWE-22', 'CWE-352', 'CWE-79'],
-                            pattern: '^CWE-[1-9]\\d{0,5}$',
-                            metaInfo: {},
-                            type: 'STRING',
+                      {
+                        key: 'name',
+                        fullName: ['vulnerabilities', 'cwe', 'name'],
+                        title: 'Weakness name',
+                        description:
+                          'Holds the full name of the weakness as given in the CWE specification.',
+                        metaData: {
+                          userDocumentation: {
+                            specification:
+                              'docs/user/vulnerabilities/vulnerability/cwe/name-spec.en.md',
+                            usage:
+                              'docs/user/vulnerabilities/vulnerability/cwe/name-usage.en.md',
                           },
-                          {
-                            key: 'name',
-                            fullName: ['vulnerabilities', 'cwes', 'name'],
-                            title: 'Weakness name',
-                            description:
-                              'Holds the full name of the weakness as given in the CWE specification.',
-                            metaData: {
-                              userDocumentation: {
-                                specification:
-                                  'docs/user/vulnerabilities/vulnerability/cwes/name-spec.en.md',
-                                usage:
-                                  'docs/user/vulnerabilities/vulnerability/cwes/name-usage.en.md',
-                              },
-                              relevanceLevels: {
-                                csaf_base: 'mandatory',
-                                csaf_security_incident_response: 'mandatory',
-                                csaf_informational_advisory: 'mandatory',
-                                csaf_security_advisory: 'mandatory',
-                                csaf_vex: 'mandatory',
-                              },
-                              i18n: {
-                                title: 'VulnerabilitiesItemsCwesNameTitle',
-                                description:
-                                  'VulnerabilitiesItemsCwesNameDescription',
-                              },
-                            },
-                            minLength: 1,
-                            examples: [
-                              'Cross-Site Request Forgery (CSRF)',
-                              "Improper Limitation of a Pathname to a Restricted Directory ('Path Traversal')",
-                              "Improper Neutralization of Input During Web Page Generation ('Cross-site Scripting')",
-                            ],
-                            metaInfo: {},
-                            type: 'STRING',
+                          relevanceLevels: {
+                            csaf_base: 'mandatory',
+                            csaf_security_incident_response: 'mandatory',
+                            csaf_informational_advisory: 'mandatory',
+                            csaf_security_advisory: 'mandatory',
+                            csaf_vex: 'mandatory',
                           },
-                          {
-                            key: 'version',
-                            fullName: ['vulnerabilities', 'cwes', 'version'],
-                            title: 'CWE version',
+                          i18n: {
+                            title: 'VulnerabilitiesItemsCweNameTitle',
                             description:
-                              'Holds the version string of the CWE specification this weakness was extracted from.',
-                            metaData: {
-                              userDocumentation: {
-                                specification:
-                                  'docs/user/vulnerabilities/vulnerability/cwes/version-spec.en.md',
-                                usage:
-                                  'docs/user/vulnerabilities/vulnerability/cwes/version-usage.en.md',
-                              },
-                              relevanceLevels: {
-                                csaf_base: 'mandatory',
-                                csaf_security_incident_response: 'mandatory',
-                                csaf_informational_advisory: 'mandatory',
-                                csaf_security_advisory: 'mandatory',
-                                csaf_vex: 'mandatory',
-                              },
-                              i18n: {
-                                title: 'VulnerabilitiesItemsCwesVersionTitle',
-                                description:
-                                  'VulnerabilitiesItemsCwesVersionDescription',
-                              },
-                            },
-                            minLength: 1,
-                            examples: ['1.0', '3.4.1', '4.0', '4.11', '4.12'],
-                            pattern:
-                              '^[1-9]\\d*\\.([0-9]|([1-9]\\d+))(\\.\\d+)?$',
-                            metaInfo: {},
-                            type: 'STRING',
+                              'VulnerabilitiesItemsCweNameDescription',
                           },
+                        },
+                        minLength: 1,
+                        examples: [
+                          'Cross-Site Request Forgery (CSRF)',
+                          "Improper Limitation of a Pathname to a Restricted Directory ('Path Traversal')",
+                          "Improper Neutralization of Input During Web Page Generation ('Cross-site Scripting')",
                         ],
+                        metaInfo: {},
+                        type: 'STRING',
                       },
-                    },
+                    ],
                   },
-                },
-                {
-                  key: 'disclosure_date',
-                  fullName: ['vulnerabilities', 'disclosure_date'],
-                  title: 'Disclosure date',
-                  description:
-                    'Holds the date and time the vulnerability was originally disclosed to the public.',
-                  metaData: {
-                    uiType: 'STRING_DATETIME',
-                    userDocumentation: {
-                      specification:
-                        'docs/user/vulnerabilities/vulnerability/disclosure_date-spec.en.md',
-                      usage:
-                        'docs/user/vulnerabilities/vulnerability/disclosure_date-usage.en.md',
-                    },
-                    relevanceLevels: {
-                      csaf_base: 'optional',
-                      csaf_security_incident_response: 'nice_to_know',
-                      csaf_informational_advisory: 'excluded',
-                      csaf_security_advisory: 'nice_to_know',
-                      csaf_vex: 'nice_to_know',
-                    },
-                    i18n: {
-                      title: 'VulnerabilitiesItemsDisclosureDateTitle',
-                      description:
-                        'VulnerabilitiesItemsDisclosureDateDescription',
-                    },
-                  },
-                  metaInfo: {},
-                  type: 'STRING',
                 },
                 {
                   key: 'discovery_date',
@@ -7106,1100 +6757,6 @@ export default /** @type {const} */ ({
                               },
                             },
                             minLength: 1,
-                            metaInfo: {},
-                            type: 'STRING',
-                          },
-                        ],
-                      },
-                    },
-                  },
-                },
-                {
-                  key: 'metrics',
-                  fullName: ['vulnerabilities', 'metrics'],
-                  title: 'List of metrics',
-                  description:
-                    'Contains metric objects for the current vulnerability.',
-                  metaData: {
-                    userDocumentation: {
-                      specification:
-                        'docs/user/vulnerabilities/vulnerability/scores-spec.en.md',
-                      usage:
-                        'docs/user/vulnerabilities/vulnerability/scores-usage.en.md',
-                    },
-                    relevanceLevels: {
-                      csaf_base: 'nice_to_know',
-                      csaf_security_incident_response: 'nice_to_know',
-                      csaf_informational_advisory: 'excluded',
-                      csaf_security_advisory: 'best_practice',
-                      csaf_vex: 'mandatory',
-                    },
-                    i18n: {
-                      title: 'VulnerabilitiesItemsScoresTitle',
-                      description: 'VulnerabilitiesItemsScoresDescription',
-                    },
-                  },
-                  type: 'ARRAY',
-                  metaInfo: {
-                    arrayType: {
-                      key: '',
-                      fullName: ['vulnerabilities', 'metrics'],
-                      title: 'metric',
-                      description:
-                        'Contains all metadata about the metric including products it applies to and the source and the content itself.',
-                      metaData: {
-                        userDocumentation: {
-                          specification:
-                            'docs/user/vulnerabilities/vulnerability/scores/score-spec.en.md',
-                          usage:
-                            'docs/user/vulnerabilities/vulnerability/scores/score-usage.en.md',
-                        },
-                        i18n: {
-                          title: 'VulnerabilitiesItemsScoresItemsTitle',
-                          description:
-                            'VulnerabilitiesItemsScoresItemsDescription',
-                        },
-                        itemName: {
-                          itemNameTranslationKey: 'arrays.scoresItemName',
-                        },
-                        relevanceLevels: {
-                          csaf_base: 'nice_to_know',
-                          csaf_security_incident_response: 'nice_to_know',
-                          csaf_informational_advisory: 'excluded',
-                          csaf_security_advisory: 'best_practice',
-                          csaf_vex: 'mandatory',
-                        },
-                      },
-                      type: 'OBJECT',
-                      metaInfo: {
-                        propertyList: [
-                          {
-                            key: 'content',
-                            fullName: ['vulnerabilities', 'metrics', 'content'],
-                            title: 'Content',
-                            description:
-                              'Specifies information about (at least one) metric or score for the given products regarding the current vulnerability.',
-                            metaData: {
-                              userDocumentation: {
-                                specification:
-                                  'docs/user/vulnerabilities/vulnerability/metrics/content-spec.en.md',
-                                usage:
-                                  'docs/user/vulnerabilities/vulnerability/metrics/content-usage.en.md',
-                              },
-                              relevanceLevels: {
-                                csaf_base: 'mandatory',
-                                csaf_security_incident_response: 'mandatory',
-                                csaf_informational_advisory: 'excluded',
-                                csaf_security_advisory: 'mandatory',
-                                csaf_vex: 'mandatory',
-                              },
-                              i18n: {
-                                title:
-                                  'VulnerabilitiesItemsMetricsContentTitle',
-                                description:
-                                  'CvssV2VersionMetricsContentDescription',
-                              },
-                            },
-                            type: 'OBJECT',
-                            metaInfo: {
-                              propertyList: [
-                                {
-                                  key: 'cvss_v2',
-                                  fullName: [
-                                    'vulnerabilities',
-                                    'metrics',
-                                    'content',
-                                    'cvss_v2',
-                                  ],
-                                  title:
-                                    'JSON Schema for Common Vulnerability Scoring System version 2.0',
-                                  metaData: {
-                                    title: 'CVSS v2',
-                                    description:
-                                      'JSON Schema for Common Vulnerability Scoring System version 2.0',
-                                    userDocumentation: {
-                                      specification:
-                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
-                                      usage:
-                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
-                                    },
-                                    relevanceLevels: {
-                                      csaf_base: 'optional',
-                                      csaf_security_incident_response:
-                                        'optional',
-                                      csaf_informational_advisory: 'excluded',
-                                      csaf_security_advisory: 'optional',
-                                      csaf_vex: 'optional',
-                                    },
-                                    i18n: {
-                                      title: 'CvssV2Title',
-                                      description: 'CvssV2Description',
-                                    },
-                                    uiType: 'OBJECT_CVSS_2',
-                                  },
-                                  type: 'OBJECT',
-                                  metaInfo: {
-                                    propertyList: [
-                                      {
-                                        key: 'version',
-                                        fullName: [
-                                          'vulnerabilities',
-                                          'metrics',
-                                          'content',
-                                          'cvss_v2',
-                                          'version',
-                                        ],
-                                        description: 'CVSS Version',
-                                        metaData: {
-                                          uiType: 'STRING_ENUM',
-                                          userDocumentation: {
-                                            specification:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
-                                            usage:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
-                                          },
-                                          relevanceLevels: {
-                                            csaf_base: 'mandatory',
-                                            csaf_security_incident_response:
-                                              'mandatory',
-                                            csaf_informational_advisory:
-                                              'excluded',
-                                            csaf_security_advisory: 'mandatory',
-                                            csaf_vex: 'mandatory',
-                                          },
-                                          i18n: {
-                                            title: 'CvssV2VersionTitle',
-                                            description:
-                                              'CvssV2VersionDescription',
-                                          },
-                                        },
-                                        enum: ['2.0'],
-                                        metaInfo: {},
-                                        type: 'STRING',
-                                      },
-                                      {
-                                        key: 'vectorString',
-                                        fullName: [
-                                          'vulnerabilities',
-                                          'metrics',
-                                          'content',
-                                          'cvss_v2',
-                                          'vectorString',
-                                        ],
-                                        metaData: {
-                                          userDocumentation: {
-                                            specification:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
-                                            usage:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
-                                          },
-                                          relevanceLevels: {
-                                            csaf_base: 'mandatory',
-                                            csaf_security_incident_response:
-                                              'mandatory',
-                                            csaf_informational_advisory:
-                                              'excluded',
-                                            csaf_security_advisory: 'mandatory',
-                                            csaf_vex: 'mandatory',
-                                          },
-                                          i18n: {
-                                            title: 'CvssV2VectorStringTitle',
-                                            description:
-                                              'CvssV2VectorStringDescription',
-                                          },
-                                        },
-                                        pattern:
-                                          '^((AV:[NAL]|AC:[LMH]|Au:[MSN]|[CIA]:[NPC]|E:(U|POC|F|H|ND)|RL:(OF|TF|W|U|ND)|RC:(UC|UR|C|ND)|CDP:(N|L|LM|MH|H|ND)|TD:(N|L|M|H|ND)|[CIA]R:(L|M|H|ND))/)*(AV:[NAL]|AC:[LMH]|Au:[MSN]|[CIA]:[NPC]|E:(U|POC|F|H|ND)|RL:(OF|TF|W|U|ND)|RC:(UC|UR|C|ND)|CDP:(N|L|LM|MH|H|ND)|TD:(N|L|M|H|ND)|[CIA]R:(L|M|H|ND))$',
-                                        metaInfo: {},
-                                        type: 'STRING',
-                                      },
-                                      {
-                                        key: 'accessVector',
-                                        fullName: [
-                                          'vulnerabilities',
-                                          'metrics',
-                                          'content',
-                                          'cvss_v2',
-                                          'accessVector',
-                                        ],
-                                        metaData: {
-                                          uiType: 'STRING_ENUM',
-                                          userDocumentation: {
-                                            specification:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
-                                            usage:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
-                                          },
-                                          relevanceLevels: {
-                                            csaf_base: 'best_practice',
-                                            csaf_security_incident_response:
-                                              'best_practice',
-                                            csaf_informational_advisory:
-                                              'excluded',
-                                            csaf_security_advisory:
-                                              'best_practice',
-                                            csaf_vex: 'best_practice',
-                                          },
-                                          i18n: {
-                                            title: 'CvssV2AccessVectorTitle',
-                                            description:
-                                              'CvssV2AccessVectorDescription',
-                                          },
-                                        },
-                                        enum: [
-                                          'NETWORK',
-                                          'ADJACENT_NETWORK',
-                                          'LOCAL',
-                                        ],
-                                        metaInfo: {},
-                                        type: 'STRING',
-                                      },
-                                      {
-                                        key: 'accessComplexity',
-                                        fullName: [
-                                          'vulnerabilities',
-                                          'metrics',
-                                          'content',
-                                          'cvss_v2',
-                                          'accessComplexity',
-                                        ],
-                                        metaData: {
-                                          uiType: 'STRING_ENUM',
-                                          userDocumentation: {
-                                            specification:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
-                                            usage:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
-                                          },
-                                          relevanceLevels: {
-                                            csaf_base: 'best_practice',
-                                            csaf_security_incident_response:
-                                              'best_practice',
-                                            csaf_informational_advisory:
-                                              'excluded',
-                                            csaf_security_advisory:
-                                              'best_practice',
-                                            csaf_vex: 'best_practice',
-                                          },
-                                          i18n: {
-                                            title:
-                                              'CvssV2AccessComplexityTitle',
-                                            description:
-                                              'CvssV2AccessComplexityDescription',
-                                          },
-                                        },
-                                        enum: ['HIGH', 'MEDIUM', 'LOW'],
-                                        metaInfo: {},
-                                        type: 'STRING',
-                                      },
-                                      {
-                                        key: 'authentication',
-                                        fullName: [
-                                          'vulnerabilities',
-                                          'metrics',
-                                          'content',
-                                          'cvss_v2',
-                                          'authentication',
-                                        ],
-                                        metaData: {
-                                          uiType: 'STRING_ENUM',
-                                          userDocumentation: {
-                                            specification:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
-                                            usage:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
-                                          },
-                                          relevanceLevels: {
-                                            csaf_base: 'best_practice',
-                                            csaf_security_incident_response:
-                                              'best_practice',
-                                            csaf_informational_advisory:
-                                              'excluded',
-                                            csaf_security_advisory:
-                                              'best_practice',
-                                            csaf_vex: 'best_practice',
-                                          },
-                                          i18n: {
-                                            title: 'CvssV2AuthenticationTitle',
-                                            description:
-                                              'CvssV2AuthenticationDescription',
-                                          },
-                                        },
-                                        enum: ['MULTIPLE', 'SINGLE', 'NONE'],
-                                        metaInfo: {},
-                                        type: 'STRING',
-                                      },
-                                      {
-                                        key: 'confidentialityImpact',
-                                        fullName: [
-                                          'vulnerabilities',
-                                          'metrics',
-                                          'content',
-                                          'cvss_v2',
-                                          'confidentialityImpact',
-                                        ],
-                                        metaData: {
-                                          uiType: 'STRING_ENUM',
-                                          userDocumentation: {
-                                            specification:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
-                                            usage:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
-                                          },
-                                          relevanceLevels: {
-                                            csaf_base: 'best_practice',
-                                            csaf_security_incident_response:
-                                              'best_practice',
-                                            csaf_informational_advisory:
-                                              'excluded',
-                                            csaf_security_advisory:
-                                              'best_practice',
-                                            csaf_vex: 'best_practice',
-                                          },
-                                          i18n: {
-                                            title:
-                                              'CvssV2ConfidentialityImpactTitle',
-                                            description:
-                                              'CvssV2ConfidentialityImpactDescription',
-                                          },
-                                        },
-                                        enum: ['NONE', 'PARTIAL', 'COMPLETE'],
-                                        metaInfo: {},
-                                        type: 'STRING',
-                                      },
-                                      {
-                                        key: 'integrityImpact',
-                                        fullName: [
-                                          'vulnerabilities',
-                                          'metrics',
-                                          'content',
-                                          'cvss_v2',
-                                          'integrityImpact',
-                                        ],
-                                        metaData: {
-                                          uiType: 'STRING_ENUM',
-                                          userDocumentation: {
-                                            specification:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
-                                            usage:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
-                                          },
-                                          relevanceLevels: {
-                                            csaf_base: 'best_practice',
-                                            csaf_security_incident_response:
-                                              'best_practice',
-                                            csaf_informational_advisory:
-                                              'excluded',
-                                            csaf_security_advisory:
-                                              'best_practice',
-                                            csaf_vex: 'best_practice',
-                                          },
-                                          i18n: {
-                                            title: 'CvssV2IntegrityImpactTitle',
-                                            description:
-                                              'CvssV2IntegrityImpactDescription',
-                                          },
-                                        },
-                                        enum: ['NONE', 'PARTIAL', 'COMPLETE'],
-                                        metaInfo: {},
-                                        type: 'STRING',
-                                      },
-                                      {
-                                        key: 'availabilityImpact',
-                                        fullName: [
-                                          'vulnerabilities',
-                                          'metrics',
-                                          'content',
-                                          'cvss_v2',
-                                          'availabilityImpact',
-                                        ],
-                                        metaData: {
-                                          uiType: 'STRING_ENUM',
-                                          userDocumentation: {
-                                            specification:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
-                                            usage:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
-                                          },
-                                          relevanceLevels: {
-                                            csaf_base: 'best_practice',
-                                            csaf_security_incident_response:
-                                              'best_practice',
-                                            csaf_informational_advisory:
-                                              'excluded',
-                                            csaf_security_advisory:
-                                              'best_practice',
-                                            csaf_vex: 'best_practice',
-                                          },
-                                          i18n: {
-                                            title:
-                                              'CvssV2AvailabilityImpactTitle',
-                                            description:
-                                              'CvssV2AvailabilityImpactDescription',
-                                          },
-                                        },
-                                        enum: ['NONE', 'PARTIAL', 'COMPLETE'],
-                                        metaInfo: {},
-                                        type: 'STRING',
-                                      },
-                                      {
-                                        key: 'baseScore',
-                                        fullName: [
-                                          'vulnerabilities',
-                                          'metrics',
-                                          'content',
-                                          'cvss_v2',
-                                          'baseScore',
-                                        ],
-                                        metaData: {
-                                          userDocumentation: {
-                                            specification:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
-                                            usage:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
-                                          },
-                                          relevanceLevels: {
-                                            csaf_base: 'mandatory',
-                                            csaf_security_incident_response:
-                                              'mandatory',
-                                            csaf_informational_advisory:
-                                              'excluded',
-                                            csaf_security_advisory: 'mandatory',
-                                            csaf_vex: 'mandatory',
-                                          },
-                                          i18n: {
-                                            title: 'CvssV2BaseScoreTitle',
-                                            description:
-                                              'CvssV2BaseScoreDescription',
-                                          },
-                                        },
-                                        metaInfo: {},
-                                        type: 'NUMBER',
-                                      },
-                                      {
-                                        key: 'exploitability',
-                                        fullName: [
-                                          'vulnerabilities',
-                                          'metrics',
-                                          'content',
-                                          'cvss_v2',
-                                          'exploitability',
-                                        ],
-                                        metaData: {
-                                          uiType: 'STRING_ENUM',
-                                          userDocumentation: {
-                                            specification:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
-                                            usage:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
-                                          },
-                                          relevanceLevels: {
-                                            csaf_base: 'nice_to_know',
-                                            csaf_security_incident_response:
-                                              'nice_to_know',
-                                            csaf_informational_advisory:
-                                              'excluded',
-                                            csaf_security_advisory:
-                                              'nice_to_know',
-                                            csaf_vex: 'nice_to_know',
-                                          },
-                                          i18n: {
-                                            title: 'CvssV2ExploitabilityTitle',
-                                            description:
-                                              'CvssV2ExploitabilityDescription',
-                                          },
-                                        },
-                                        enum: [
-                                          'UNPROVEN',
-                                          'PROOF_OF_CONCEPT',
-                                          'FUNCTIONAL',
-                                          'HIGH',
-                                          'NOT_DEFINED',
-                                        ],
-                                        metaInfo: {},
-                                        type: 'STRING',
-                                      },
-                                      {
-                                        key: 'remediationLevel',
-                                        fullName: [
-                                          'vulnerabilities',
-                                          'metrics',
-                                          'content',
-                                          'cvss_v2',
-                                          'remediationLevel',
-                                        ],
-                                        metaData: {
-                                          uiType: 'STRING_ENUM',
-                                          userDocumentation: {
-                                            specification:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
-                                            usage:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
-                                          },
-                                          relevanceLevels: {
-                                            csaf_base: 'want_to_have',
-                                            csaf_security_incident_response:
-                                              'want_to_have',
-                                            csaf_informational_advisory:
-                                              'excluded',
-                                            csaf_security_advisory:
-                                              'want_to_have',
-                                            csaf_vex: 'want_to_have',
-                                          },
-                                          i18n: {
-                                            title:
-                                              'CvssV2RemediationLevelTitle',
-                                            description:
-                                              'CvssV2RemediationLevelDescription',
-                                          },
-                                        },
-                                        enum: [
-                                          'OFFICIAL_FIX',
-                                          'TEMPORARY_FIX',
-                                          'WORKAROUND',
-                                          'UNAVAILABLE',
-                                          'NOT_DEFINED',
-                                        ],
-                                        metaInfo: {},
-                                        type: 'STRING',
-                                      },
-                                      {
-                                        key: 'reportConfidence',
-                                        fullName: [
-                                          'vulnerabilities',
-                                          'metrics',
-                                          'content',
-                                          'cvss_v2',
-                                          'reportConfidence',
-                                        ],
-                                        metaData: {
-                                          uiType: 'STRING_ENUM',
-                                          userDocumentation: {
-                                            specification:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
-                                            usage:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
-                                          },
-                                          relevanceLevels: {
-                                            csaf_base: 'want_to_have',
-                                            csaf_security_incident_response:
-                                              'want_to_have',
-                                            csaf_informational_advisory:
-                                              'excluded',
-                                            csaf_security_advisory:
-                                              'want_to_have',
-                                            csaf_vex: 'want_to_have',
-                                          },
-                                          i18n: {
-                                            title:
-                                              'CvssV2ReportConfidenceTitle',
-                                            description:
-                                              'CvssV2ReportConfidenceDescription',
-                                          },
-                                        },
-                                        enum: [
-                                          'UNCONFIRMED',
-                                          'UNCORROBORATED',
-                                          'CONFIRMED',
-                                          'NOT_DEFINED',
-                                        ],
-                                        metaInfo: {},
-                                        type: 'STRING',
-                                      },
-                                      {
-                                        key: 'temporalScore',
-                                        fullName: [
-                                          'vulnerabilities',
-                                          'metrics',
-                                          'content',
-                                          'cvss_v2',
-                                          'temporalScore',
-                                        ],
-                                        metaData: {
-                                          userDocumentation: {
-                                            specification:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
-                                            usage:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
-                                          },
-                                          relevanceLevels: {
-                                            csaf_base: 'want_to_have',
-                                            csaf_security_incident_response:
-                                              'want_to_have',
-                                            csaf_informational_advisory:
-                                              'excluded',
-                                            csaf_security_advisory:
-                                              'want_to_have',
-                                            csaf_vex: 'want_to_have',
-                                          },
-                                          i18n: {
-                                            title: 'CvssV2TemporalScoreTitle',
-                                            description:
-                                              'CvssV2TemporalScoreDescription',
-                                          },
-                                        },
-                                        metaInfo: {},
-                                        type: 'NUMBER',
-                                      },
-                                      {
-                                        key: 'collateralDamagePotential',
-                                        fullName: [
-                                          'vulnerabilities',
-                                          'metrics',
-                                          'content',
-                                          'cvss_v2',
-                                          'collateralDamagePotential',
-                                        ],
-                                        metaData: {
-                                          uiType: 'STRING_ENUM',
-                                          userDocumentation: {
-                                            specification:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
-                                            usage:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
-                                          },
-                                          relevanceLevels: {
-                                            csaf_base: 'optional',
-                                            csaf_security_incident_response:
-                                              'optional',
-                                            csaf_informational_advisory:
-                                              'excluded',
-                                            csaf_security_advisory: 'optional',
-                                            csaf_vex: 'optional',
-                                          },
-                                          i18n: {
-                                            title:
-                                              'CvssV2CollateralDamagePotentialTitle',
-                                            description:
-                                              'CvssV2CollateralDamagePotentialDescription',
-                                          },
-                                        },
-                                        enum: [
-                                          'NONE',
-                                          'LOW',
-                                          'LOW_MEDIUM',
-                                          'MEDIUM_HIGH',
-                                          'HIGH',
-                                          'NOT_DEFINED',
-                                        ],
-                                        metaInfo: {},
-                                        type: 'STRING',
-                                      },
-                                      {
-                                        key: 'targetDistribution',
-                                        fullName: [
-                                          'vulnerabilities',
-                                          'metrics',
-                                          'content',
-                                          'cvss_v2',
-                                          'targetDistribution',
-                                        ],
-                                        metaData: {
-                                          uiType: 'STRING_ENUM',
-                                          userDocumentation: {
-                                            specification:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
-                                            usage:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
-                                          },
-                                          relevanceLevels: {
-                                            csaf_base: 'optional',
-                                            csaf_security_incident_response:
-                                              'optional',
-                                            csaf_informational_advisory:
-                                              'excluded',
-                                            csaf_security_advisory: 'optional',
-                                            csaf_vex: 'optional',
-                                          },
-                                          i18n: {
-                                            title:
-                                              'CvssV2TargetDistributionTitle',
-                                            description:
-                                              'CvssV2TargetDistributionDescription',
-                                          },
-                                        },
-                                        enum: [
-                                          'NONE',
-                                          'LOW',
-                                          'MEDIUM',
-                                          'HIGH',
-                                          'NOT_DEFINED',
-                                        ],
-                                        metaInfo: {},
-                                        type: 'STRING',
-                                      },
-                                      {
-                                        key: 'confidentialityRequirement',
-                                        fullName: [
-                                          'vulnerabilities',
-                                          'metrics',
-                                          'content',
-                                          'cvss_v2',
-                                          'confidentialityRequirement',
-                                        ],
-                                        metaData: {
-                                          uiType: 'STRING_ENUM',
-                                          userDocumentation: {
-                                            specification:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
-                                            usage:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
-                                          },
-                                          relevanceLevels: {
-                                            csaf_base: 'optional',
-                                            csaf_security_incident_response:
-                                              'optional',
-                                            csaf_informational_advisory:
-                                              'excluded',
-                                            csaf_security_advisory: 'optional',
-                                            csaf_vex: 'optional',
-                                          },
-                                          i18n: {
-                                            title:
-                                              'CvssV2ConfidentialityRequirementTitle',
-                                            description:
-                                              'CvssV2ConfidentialityRequirementDescription',
-                                          },
-                                        },
-                                        enum: [
-                                          'LOW',
-                                          'MEDIUM',
-                                          'HIGH',
-                                          'NOT_DEFINED',
-                                        ],
-                                        metaInfo: {},
-                                        type: 'STRING',
-                                      },
-                                      {
-                                        key: 'integrityRequirement',
-                                        fullName: [
-                                          'vulnerabilities',
-                                          'metrics',
-                                          'content',
-                                          'cvss_v2',
-                                          'integrityRequirement',
-                                        ],
-                                        metaData: {
-                                          uiType: 'STRING_ENUM',
-                                          userDocumentation: {
-                                            specification:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
-                                            usage:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
-                                          },
-                                          relevanceLevels: {
-                                            csaf_base: 'optional',
-                                            csaf_security_incident_response:
-                                              'optional',
-                                            csaf_informational_advisory:
-                                              'excluded',
-                                            csaf_security_advisory: 'optional',
-                                            csaf_vex: 'optional',
-                                          },
-                                          i18n: {
-                                            title:
-                                              'CvssV2IntegrityRequirementTitle',
-                                            description:
-                                              'CvssV2IntegrityRequirementDescription',
-                                          },
-                                        },
-                                        enum: [
-                                          'LOW',
-                                          'MEDIUM',
-                                          'HIGH',
-                                          'NOT_DEFINED',
-                                        ],
-                                        metaInfo: {},
-                                        type: 'STRING',
-                                      },
-                                      {
-                                        key: 'availabilityRequirement',
-                                        fullName: [
-                                          'vulnerabilities',
-                                          'metrics',
-                                          'content',
-                                          'cvss_v2',
-                                          'availabilityRequirement',
-                                        ],
-                                        metaData: {
-                                          uiType: 'STRING_ENUM',
-                                          userDocumentation: {
-                                            specification:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
-                                            usage:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
-                                          },
-                                          relevanceLevels: {
-                                            csaf_base: 'optional',
-                                            csaf_security_incident_response:
-                                              'optional',
-                                            csaf_informational_advisory:
-                                              'excluded',
-                                            csaf_security_advisory: 'optional',
-                                            csaf_vex: 'optional',
-                                          },
-                                          i18n: {
-                                            title:
-                                              'CvssV2AvailabilityRequirementTitle',
-                                            description:
-                                              'CvssV2AvailabilityRequirementDescription',
-                                          },
-                                        },
-                                        enum: [
-                                          'LOW',
-                                          'MEDIUM',
-                                          'HIGH',
-                                          'NOT_DEFINED',
-                                        ],
-                                        metaInfo: {},
-                                        type: 'STRING',
-                                      },
-                                      {
-                                        key: 'environmentalScore',
-                                        fullName: [
-                                          'vulnerabilities',
-                                          'metrics',
-                                          'content',
-                                          'cvss_v2',
-                                          'environmentalScore',
-                                        ],
-                                        metaData: {
-                                          userDocumentation: {
-                                            specification:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
-                                            usage:
-                                              'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
-                                          },
-                                          relevanceLevels: {
-                                            csaf_base: 'optional',
-                                            csaf_security_incident_response:
-                                              'optional',
-                                            csaf_informational_advisory:
-                                              'excluded',
-                                            csaf_security_advisory: 'optional',
-                                            csaf_vex: 'optional',
-                                          },
-                                          i18n: {
-                                            title:
-                                              'CvssV2EnvironmentalScoreTitle',
-                                            description:
-                                              'CvssV2EnvironmentalScoreDescription',
-                                          },
-                                        },
-                                        metaInfo: {},
-                                        type: 'NUMBER',
-                                      },
-                                    ],
-                                  },
-                                },
-                                {
-                                  key: 'cvss_v3',
-                                  fullName: [
-                                    'vulnerabilities',
-                                    'metrics',
-                                    'content',
-                                    'cvss_v3',
-                                  ],
-                                  metaData: {
-                                    title: 'CVSS v3',
-                                    description:
-                                      'JSON Schema for Common Vulnerability Scoring System version 3.0',
-                                    userDocumentation: {
-                                      specification:
-                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v3-spec.en.md',
-                                      usage:
-                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v3-usage.en.md',
-                                    },
-                                    relevanceLevels: {
-                                      csaf_base: 'best_practice',
-                                      csaf_security_incident_response:
-                                        'best_practice',
-                                      csaf_informational_advisory: 'excluded',
-                                      csaf_security_advisory: 'best_practice',
-                                      csaf_vex: 'best_practice',
-                                    },
-                                    i18n: {
-                                      title: 'CvssV3Title',
-                                      description: 'CvssV3Description',
-                                    },
-                                    uiType: 'OBJECT_CVSS_3',
-                                  },
-                                  type: 'OBJECT',
-                                  metaInfo: { propertyList: [] },
-                                },
-                                {
-                                  key: 'cvss_v4',
-                                  fullName: [
-                                    'vulnerabilities',
-                                    'metrics',
-                                    'content',
-                                    'cvss_v4',
-                                  ],
-                                  metaData: {
-                                    title: 'CVSS v3',
-                                    description:
-                                      'JSON Schema for Common Vulnerability Scoring System version 3.0',
-                                    userDocumentation: {
-                                      specification:
-                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v4-spec.en.md',
-                                      usage:
-                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v4-usage.en.md',
-                                    },
-                                    relevanceLevels: {
-                                      csaf_base: 'best_practice',
-                                      csaf_security_incident_response:
-                                        'best_practice',
-                                      csaf_informational_advisory: 'excluded',
-                                      csaf_security_advisory: 'best_practice',
-                                      csaf_vex: 'best_practice',
-                                    },
-                                    i18n: {
-                                      title: 'CvssV4Title',
-                                      description: 'CvssV4Description',
-                                    },
-                                    uiType: 'OBJECT_CVSS_3',
-                                  },
-                                  type: 'OBJECT',
-                                  metaInfo: { propertyList: [] },
-                                },
-                                {
-                                  key: 'ssvc_v1',
-                                  fullName: [
-                                    'vulnerabilities',
-                                    'metrics',
-                                    'content',
-                                    'ssvc_v1',
-                                  ],
-                                  metaData: {
-                                    title: 'SSVC V1',
-                                    description:
-                                      'JSON Schema for Stakeholder-Specific Vulnerability Categorization version 1.0',
-                                    userDocumentation: {
-                                      specification:
-                                        'docs/user/vulnerabilities/vulnerability/scores/score/ssvc_v1-spec.en.md',
-                                      usage:
-                                        'docs/user/vulnerabilities/vulnerability/scores/score/ssvc_v1-usage.en.md',
-                                    },
-                                    relevanceLevels: {
-                                      csaf_base: 'best_practice',
-                                      csaf_security_incident_response:
-                                        'best_practice',
-                                      csaf_informational_advisory: 'excluded',
-                                      csaf_security_advisory: 'best_practice',
-                                      csaf_vex: 'best_practice',
-                                    },
-                                    i18n: {
-                                      title: 'SsvcV1Title',
-                                      description: 'SsvcV1Description',
-                                    },
-                                    uiType: 'OBJECT_CVSS_3',
-                                  },
-                                  type: 'OBJECT',
-                                  metaInfo: { propertyList: [] },
-                                },
-                              ],
-                            },
-                          },
-                          {
-                            key: 'products',
-                            fullName: [
-                              'vulnerabilities',
-                              'metrics',
-                              'products',
-                            ],
-                            title: 'List of product_ids',
-                            description:
-                              'Specifies a list of product_ids to give context to the parent item.',
-                            metaData: {
-                              userDocumentation: {
-                                specification:
-                                  'docs/user/vulnerabilities/vulnerability/scores/score/products-spec.en.md',
-                                usage:
-                                  'docs/user/vulnerabilities/vulnerability/scores/score/products-usage.en.md',
-                              },
-                              relevanceLevels: {
-                                csaf_base: 'mandatory',
-                                csaf_security_incident_response: 'mandatory',
-                                csaf_informational_advisory: 'mandatory',
-                                csaf_security_advisory: 'mandatory',
-                                csaf_vex: 'mandatory',
-                              },
-                              i18n: {
-                                title: 'ProductsTitle',
-                                description: 'ProductsDescription',
-                              },
-                            },
-                            type: 'ARRAY',
-                            metaInfo: {
-                              arrayType: {
-                                key: '',
-                                fullName: [
-                                  'vulnerabilities',
-                                  'metrics',
-                                  'products',
-                                ],
-                                title: 'Reference token for product instance',
-                                description:
-                                  'Token required to identify a full_product_name so that it can be referred to from other parts in the document. There is no predefined or required format for the product_id as long as it uniquely identifies a product in the context of the current document.',
-                                metaData: {
-                                  uiType: 'STRING_PRODUCT_ID',
-                                  userDocumentation: {
-                                    specification:
-                                      'docs/user/types/product_id-spec.en.md',
-                                    usage:
-                                      'docs/user/types/product_id-usage.en.md',
-                                  },
-                                  i18n: {
-                                    title: 'ProductIdTitle',
-                                    description: 'ProductIdDescription',
-                                  },
-                                  itemName: {
-                                    itemNameTranslationKey:
-                                      'arrays.productsItemName',
-                                  },
-                                  relevanceLevels: {
-                                    csaf_base: 'mandatory',
-                                    csaf_security_incident_response:
-                                      'mandatory',
-                                    csaf_informational_advisory: 'mandatory',
-                                    csaf_security_advisory: 'mandatory',
-                                    csaf_vex: 'mandatory',
-                                  },
-                                },
-                                minLength: 1,
-                                examples: ['CSAFPID-0004', 'CSAFPID-0008'],
-                                metaInfo: {},
-                                type: 'STRING',
-                              },
-                            },
-                          },
-                          {
-                            key: 'source',
-                            fullName: ['vulnerabilities', 'metrics', 'source'],
-                            title: 'Source',
-                            description:
-                              'Contains the URL of the source that originally determined the metric.',
-                            metaData: {
-                              uiType: 'STRING_URI',
-                              userDocumentation: {
-                                specification:
-                                  'docs/user/vulnerabilities/vulnerability/metrics/source-spec.en.md',
-                                usage:
-                                  'docs/user/vulnerabilities/vulnerability/metrics/source-usage.en.md',
-                              },
-                              relevanceLevels: {
-                                csaf_base: 'mandatory',
-                                csaf_security_incident_response: 'mandatory',
-                                csaf_informational_advisory: 'excluded',
-                                csaf_security_advisory: 'mandatory',
-                                csaf_vex: 'mandatory',
-                              },
-                              i18n: {
-                                title: 'VulnerabilitiesItemsMetricsSourceTitle',
-                                description:
-                                  'CvssV2VersionMetricsSourceDescription',
-                              },
-                            },
                             metaInfo: {},
                             type: 'STRING',
                           },
@@ -9188,6 +7745,35 @@ export default /** @type {const} */ ({
                   },
                 },
                 {
+                  key: 'release_date',
+                  fullName: ['vulnerabilities', 'release_date'],
+                  title: 'Release date',
+                  description:
+                    'Holds the date and time the vulnerability was originally released into the wild.',
+                  metaData: {
+                    uiType: 'STRING_DATETIME',
+                    userDocumentation: {
+                      specification:
+                        'docs/user/vulnerabilities/vulnerability/release_date-spec.en.md',
+                      usage:
+                        'docs/user/vulnerabilities/vulnerability/release_date-usage.en.md',
+                    },
+                    relevanceLevels: {
+                      csaf_base: 'optional',
+                      csaf_security_incident_response: 'nice_to_know',
+                      csaf_informational_advisory: 'excluded',
+                      csaf_security_advisory: 'nice_to_know',
+                      csaf_vex: 'nice_to_know',
+                    },
+                    i18n: {
+                      title: 'VulnerabilitiesItemsReleaseDateTitle',
+                      description: 'VulnerabilitiesItemsReleaseDateDescription',
+                    },
+                  },
+                  metaInfo: {},
+                  type: 'STRING',
+                },
+                {
                   key: 'remediations',
                   fullName: ['vulnerabilities', 'remediations'],
                   title: 'List of remediations',
@@ -9279,11 +7865,9 @@ export default /** @type {const} */ ({
                               },
                             },
                             enum: [
-                              'fix_planned',
                               'mitigation',
                               'no_fix_planned',
                               'none_available',
-                              'optional_patch',
                               'vendor_fix',
                               'workaround',
                             ],
@@ -9743,6 +8327,894 @@ export default /** @type {const} */ ({
                             },
                             metaInfo: {},
                             type: 'STRING',
+                          },
+                        ],
+                      },
+                    },
+                  },
+                },
+                {
+                  key: 'scores',
+                  fullName: ['vulnerabilities', 'scores'],
+                  title: 'List of scores',
+                  description:
+                    'Contains score objects for the current vulnerability.',
+                  metaData: {
+                    userDocumentation: {
+                      specification:
+                        'docs/user/vulnerabilities/vulnerability/scores-spec.en.md',
+                      usage:
+                        'docs/user/vulnerabilities/vulnerability/scores-usage.en.md',
+                    },
+                    relevanceLevels: {
+                      csaf_base: 'nice_to_know',
+                      csaf_security_incident_response: 'nice_to_know',
+                      csaf_informational_advisory: 'excluded',
+                      csaf_security_advisory: 'best_practice',
+                      csaf_vex: 'mandatory',
+                    },
+                    i18n: {
+                      title: 'VulnerabilitiesItemsScoresTitle',
+                      description: 'VulnerabilitiesItemsScoresDescription',
+                    },
+                  },
+                  type: 'ARRAY',
+                  metaInfo: {
+                    arrayType: {
+                      key: '',
+                      fullName: ['vulnerabilities', 'scores'],
+                      title: 'Score',
+                      description:
+                        'Specifies information about (at least one) score of the vulnerability and for which products the given value applies.',
+                      metaData: {
+                        userDocumentation: {
+                          specification:
+                            'docs/user/vulnerabilities/vulnerability/scores/score-spec.en.md',
+                          usage:
+                            'docs/user/vulnerabilities/vulnerability/scores/score-usage.en.md',
+                        },
+                        i18n: {
+                          title: 'VulnerabilitiesItemsScoresItemsTitle',
+                          description:
+                            'VulnerabilitiesItemsScoresItemsDescription',
+                        },
+                        itemName: {
+                          itemNameTranslationKey: 'arrays.scoresItemName',
+                        },
+                        relevanceLevels: {
+                          csaf_base: 'nice_to_know',
+                          csaf_security_incident_response: 'nice_to_know',
+                          csaf_informational_advisory: 'excluded',
+                          csaf_security_advisory: 'best_practice',
+                          csaf_vex: 'mandatory',
+                        },
+                      },
+                      type: 'OBJECT',
+                      metaInfo: {
+                        propertyList: [
+                          {
+                            key: 'cvss_v2',
+                            fullName: ['vulnerabilities', 'scores', 'cvss_v2'],
+                            title:
+                              'JSON Schema for Common Vulnerability Scoring System version 2.0',
+                            metaData: {
+                              title: 'CVSS v2',
+                              description:
+                                'JSON Schema for Common Vulnerability Scoring System version 2.0',
+                              userDocumentation: {
+                                specification:
+                                  'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
+                                usage:
+                                  'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
+                              },
+                              relevanceLevels: {
+                                csaf_base: 'optional',
+                                csaf_security_incident_response: 'optional',
+                                csaf_informational_advisory: 'excluded',
+                                csaf_security_advisory: 'optional',
+                                csaf_vex: 'optional',
+                              },
+                              i18n: {
+                                title: 'CvssV2Title',
+                                description: 'CvssV2Description',
+                              },
+                              uiType: 'OBJECT_CVSS_2',
+                            },
+                            type: 'OBJECT',
+                            metaInfo: {
+                              propertyList: [
+                                {
+                                  key: 'version',
+                                  fullName: [
+                                    'vulnerabilities',
+                                    'scores',
+                                    'cvss_v2',
+                                    'version',
+                                  ],
+                                  description: 'CVSS Version',
+                                  metaData: {
+                                    uiType: 'STRING_ENUM',
+                                    userDocumentation: {
+                                      specification:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
+                                      usage:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
+                                    },
+                                    relevanceLevels: {
+                                      csaf_base: 'mandatory',
+                                      csaf_security_incident_response:
+                                        'mandatory',
+                                      csaf_informational_advisory: 'excluded',
+                                      csaf_security_advisory: 'mandatory',
+                                      csaf_vex: 'mandatory',
+                                    },
+                                    i18n: {
+                                      title: 'CvssV2VersionTitle',
+                                      description: 'CvssV2VersionDescription',
+                                    },
+                                  },
+                                  enum: ['2.0'],
+                                  metaInfo: {},
+                                  type: 'STRING',
+                                },
+                                {
+                                  key: 'vectorString',
+                                  fullName: [
+                                    'vulnerabilities',
+                                    'scores',
+                                    'cvss_v2',
+                                    'vectorString',
+                                  ],
+                                  metaData: {
+                                    userDocumentation: {
+                                      specification:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
+                                      usage:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
+                                    },
+                                    relevanceLevels: {
+                                      csaf_base: 'mandatory',
+                                      csaf_security_incident_response:
+                                        'mandatory',
+                                      csaf_informational_advisory: 'excluded',
+                                      csaf_security_advisory: 'mandatory',
+                                      csaf_vex: 'mandatory',
+                                    },
+                                    i18n: {
+                                      title: 'CvssV2VectorStringTitle',
+                                      description:
+                                        'CvssV2VectorStringDescription',
+                                    },
+                                  },
+                                  pattern:
+                                    '^((AV:[NAL]|AC:[LMH]|Au:[MSN]|[CIA]:[NPC]|E:(U|POC|F|H|ND)|RL:(OF|TF|W|U|ND)|RC:(UC|UR|C|ND)|CDP:(N|L|LM|MH|H|ND)|TD:(N|L|M|H|ND)|[CIA]R:(L|M|H|ND))/)*(AV:[NAL]|AC:[LMH]|Au:[MSN]|[CIA]:[NPC]|E:(U|POC|F|H|ND)|RL:(OF|TF|W|U|ND)|RC:(UC|UR|C|ND)|CDP:(N|L|LM|MH|H|ND)|TD:(N|L|M|H|ND)|[CIA]R:(L|M|H|ND))$',
+                                  metaInfo: {},
+                                  type: 'STRING',
+                                },
+                                {
+                                  key: 'accessVector',
+                                  fullName: [
+                                    'vulnerabilities',
+                                    'scores',
+                                    'cvss_v2',
+                                    'accessVector',
+                                  ],
+                                  metaData: {
+                                    uiType: 'STRING_ENUM',
+                                    userDocumentation: {
+                                      specification:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
+                                      usage:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
+                                    },
+                                    relevanceLevels: {
+                                      csaf_base: 'best_practice',
+                                      csaf_security_incident_response:
+                                        'best_practice',
+                                      csaf_informational_advisory: 'excluded',
+                                      csaf_security_advisory: 'best_practice',
+                                      csaf_vex: 'best_practice',
+                                    },
+                                    i18n: {
+                                      title: 'CvssV2AccessVectorTitle',
+                                      description:
+                                        'CvssV2AccessVectorDescription',
+                                    },
+                                  },
+                                  enum: [
+                                    'NETWORK',
+                                    'ADJACENT_NETWORK',
+                                    'LOCAL',
+                                  ],
+                                  metaInfo: {},
+                                  type: 'STRING',
+                                },
+                                {
+                                  key: 'accessComplexity',
+                                  fullName: [
+                                    'vulnerabilities',
+                                    'scores',
+                                    'cvss_v2',
+                                    'accessComplexity',
+                                  ],
+                                  metaData: {
+                                    uiType: 'STRING_ENUM',
+                                    userDocumentation: {
+                                      specification:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
+                                      usage:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
+                                    },
+                                    relevanceLevels: {
+                                      csaf_base: 'best_practice',
+                                      csaf_security_incident_response:
+                                        'best_practice',
+                                      csaf_informational_advisory: 'excluded',
+                                      csaf_security_advisory: 'best_practice',
+                                      csaf_vex: 'best_practice',
+                                    },
+                                    i18n: {
+                                      title: 'CvssV2AccessComplexityTitle',
+                                      description:
+                                        'CvssV2AccessComplexityDescription',
+                                    },
+                                  },
+                                  enum: ['HIGH', 'MEDIUM', 'LOW'],
+                                  metaInfo: {},
+                                  type: 'STRING',
+                                },
+                                {
+                                  key: 'authentication',
+                                  fullName: [
+                                    'vulnerabilities',
+                                    'scores',
+                                    'cvss_v2',
+                                    'authentication',
+                                  ],
+                                  metaData: {
+                                    uiType: 'STRING_ENUM',
+                                    userDocumentation: {
+                                      specification:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
+                                      usage:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
+                                    },
+                                    relevanceLevels: {
+                                      csaf_base: 'best_practice',
+                                      csaf_security_incident_response:
+                                        'best_practice',
+                                      csaf_informational_advisory: 'excluded',
+                                      csaf_security_advisory: 'best_practice',
+                                      csaf_vex: 'best_practice',
+                                    },
+                                    i18n: {
+                                      title: 'CvssV2AuthenticationTitle',
+                                      description:
+                                        'CvssV2AuthenticationDescription',
+                                    },
+                                  },
+                                  enum: ['MULTIPLE', 'SINGLE', 'NONE'],
+                                  metaInfo: {},
+                                  type: 'STRING',
+                                },
+                                {
+                                  key: 'confidentialityImpact',
+                                  fullName: [
+                                    'vulnerabilities',
+                                    'scores',
+                                    'cvss_v2',
+                                    'confidentialityImpact',
+                                  ],
+                                  metaData: {
+                                    uiType: 'STRING_ENUM',
+                                    userDocumentation: {
+                                      specification:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
+                                      usage:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
+                                    },
+                                    relevanceLevels: {
+                                      csaf_base: 'best_practice',
+                                      csaf_security_incident_response:
+                                        'best_practice',
+                                      csaf_informational_advisory: 'excluded',
+                                      csaf_security_advisory: 'best_practice',
+                                      csaf_vex: 'best_practice',
+                                    },
+                                    i18n: {
+                                      title: 'CvssV2ConfidentialityImpactTitle',
+                                      description:
+                                        'CvssV2ConfidentialityImpactDescription',
+                                    },
+                                  },
+                                  enum: ['NONE', 'PARTIAL', 'COMPLETE'],
+                                  metaInfo: {},
+                                  type: 'STRING',
+                                },
+                                {
+                                  key: 'integrityImpact',
+                                  fullName: [
+                                    'vulnerabilities',
+                                    'scores',
+                                    'cvss_v2',
+                                    'integrityImpact',
+                                  ],
+                                  metaData: {
+                                    uiType: 'STRING_ENUM',
+                                    userDocumentation: {
+                                      specification:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
+                                      usage:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
+                                    },
+                                    relevanceLevels: {
+                                      csaf_base: 'best_practice',
+                                      csaf_security_incident_response:
+                                        'best_practice',
+                                      csaf_informational_advisory: 'excluded',
+                                      csaf_security_advisory: 'best_practice',
+                                      csaf_vex: 'best_practice',
+                                    },
+                                    i18n: {
+                                      title: 'CvssV2IntegrityImpactTitle',
+                                      description:
+                                        'CvssV2IntegrityImpactDescription',
+                                    },
+                                  },
+                                  enum: ['NONE', 'PARTIAL', 'COMPLETE'],
+                                  metaInfo: {},
+                                  type: 'STRING',
+                                },
+                                {
+                                  key: 'availabilityImpact',
+                                  fullName: [
+                                    'vulnerabilities',
+                                    'scores',
+                                    'cvss_v2',
+                                    'availabilityImpact',
+                                  ],
+                                  metaData: {
+                                    uiType: 'STRING_ENUM',
+                                    userDocumentation: {
+                                      specification:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
+                                      usage:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
+                                    },
+                                    relevanceLevels: {
+                                      csaf_base: 'best_practice',
+                                      csaf_security_incident_response:
+                                        'best_practice',
+                                      csaf_informational_advisory: 'excluded',
+                                      csaf_security_advisory: 'best_practice',
+                                      csaf_vex: 'best_practice',
+                                    },
+                                    i18n: {
+                                      title: 'CvssV2AvailabilityImpactTitle',
+                                      description:
+                                        'CvssV2AvailabilityImpactDescription',
+                                    },
+                                  },
+                                  enum: ['NONE', 'PARTIAL', 'COMPLETE'],
+                                  metaInfo: {},
+                                  type: 'STRING',
+                                },
+                                {
+                                  key: 'baseScore',
+                                  fullName: [
+                                    'vulnerabilities',
+                                    'scores',
+                                    'cvss_v2',
+                                    'baseScore',
+                                  ],
+                                  metaData: {
+                                    userDocumentation: {
+                                      specification:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
+                                      usage:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
+                                    },
+                                    relevanceLevels: {
+                                      csaf_base: 'mandatory',
+                                      csaf_security_incident_response:
+                                        'mandatory',
+                                      csaf_informational_advisory: 'excluded',
+                                      csaf_security_advisory: 'mandatory',
+                                      csaf_vex: 'mandatory',
+                                    },
+                                    i18n: {
+                                      title: 'CvssV2BaseScoreTitle',
+                                      description: 'CvssV2BaseScoreDescription',
+                                    },
+                                  },
+                                  metaInfo: {},
+                                  type: 'NUMBER',
+                                },
+                                {
+                                  key: 'exploitability',
+                                  fullName: [
+                                    'vulnerabilities',
+                                    'scores',
+                                    'cvss_v2',
+                                    'exploitability',
+                                  ],
+                                  metaData: {
+                                    uiType: 'STRING_ENUM',
+                                    userDocumentation: {
+                                      specification:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
+                                      usage:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
+                                    },
+                                    relevanceLevels: {
+                                      csaf_base: 'nice_to_know',
+                                      csaf_security_incident_response:
+                                        'nice_to_know',
+                                      csaf_informational_advisory: 'excluded',
+                                      csaf_security_advisory: 'nice_to_know',
+                                      csaf_vex: 'nice_to_know',
+                                    },
+                                    i18n: {
+                                      title: 'CvssV2ExploitabilityTitle',
+                                      description:
+                                        'CvssV2ExploitabilityDescription',
+                                    },
+                                  },
+                                  enum: [
+                                    'UNPROVEN',
+                                    'PROOF_OF_CONCEPT',
+                                    'FUNCTIONAL',
+                                    'HIGH',
+                                    'NOT_DEFINED',
+                                  ],
+                                  metaInfo: {},
+                                  type: 'STRING',
+                                },
+                                {
+                                  key: 'remediationLevel',
+                                  fullName: [
+                                    'vulnerabilities',
+                                    'scores',
+                                    'cvss_v2',
+                                    'remediationLevel',
+                                  ],
+                                  metaData: {
+                                    uiType: 'STRING_ENUM',
+                                    userDocumentation: {
+                                      specification:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
+                                      usage:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
+                                    },
+                                    relevanceLevels: {
+                                      csaf_base: 'want_to_have',
+                                      csaf_security_incident_response:
+                                        'want_to_have',
+                                      csaf_informational_advisory: 'excluded',
+                                      csaf_security_advisory: 'want_to_have',
+                                      csaf_vex: 'want_to_have',
+                                    },
+                                    i18n: {
+                                      title: 'CvssV2RemediationLevelTitle',
+                                      description:
+                                        'CvssV2RemediationLevelDescription',
+                                    },
+                                  },
+                                  enum: [
+                                    'OFFICIAL_FIX',
+                                    'TEMPORARY_FIX',
+                                    'WORKAROUND',
+                                    'UNAVAILABLE',
+                                    'NOT_DEFINED',
+                                  ],
+                                  metaInfo: {},
+                                  type: 'STRING',
+                                },
+                                {
+                                  key: 'reportConfidence',
+                                  fullName: [
+                                    'vulnerabilities',
+                                    'scores',
+                                    'cvss_v2',
+                                    'reportConfidence',
+                                  ],
+                                  metaData: {
+                                    uiType: 'STRING_ENUM',
+                                    userDocumentation: {
+                                      specification:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
+                                      usage:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
+                                    },
+                                    relevanceLevels: {
+                                      csaf_base: 'want_to_have',
+                                      csaf_security_incident_response:
+                                        'want_to_have',
+                                      csaf_informational_advisory: 'excluded',
+                                      csaf_security_advisory: 'want_to_have',
+                                      csaf_vex: 'want_to_have',
+                                    },
+                                    i18n: {
+                                      title: 'CvssV2ReportConfidenceTitle',
+                                      description:
+                                        'CvssV2ReportConfidenceDescription',
+                                    },
+                                  },
+                                  enum: [
+                                    'UNCONFIRMED',
+                                    'UNCORROBORATED',
+                                    'CONFIRMED',
+                                    'NOT_DEFINED',
+                                  ],
+                                  metaInfo: {},
+                                  type: 'STRING',
+                                },
+                                {
+                                  key: 'temporalScore',
+                                  fullName: [
+                                    'vulnerabilities',
+                                    'scores',
+                                    'cvss_v2',
+                                    'temporalScore',
+                                  ],
+                                  metaData: {
+                                    userDocumentation: {
+                                      specification:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
+                                      usage:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
+                                    },
+                                    relevanceLevels: {
+                                      csaf_base: 'want_to_have',
+                                      csaf_security_incident_response:
+                                        'want_to_have',
+                                      csaf_informational_advisory: 'excluded',
+                                      csaf_security_advisory: 'want_to_have',
+                                      csaf_vex: 'want_to_have',
+                                    },
+                                    i18n: {
+                                      title: 'CvssV2TemporalScoreTitle',
+                                      description:
+                                        'CvssV2TemporalScoreDescription',
+                                    },
+                                  },
+                                  metaInfo: {},
+                                  type: 'NUMBER',
+                                },
+                                {
+                                  key: 'collateralDamagePotential',
+                                  fullName: [
+                                    'vulnerabilities',
+                                    'scores',
+                                    'cvss_v2',
+                                    'collateralDamagePotential',
+                                  ],
+                                  metaData: {
+                                    uiType: 'STRING_ENUM',
+                                    userDocumentation: {
+                                      specification:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
+                                      usage:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
+                                    },
+                                    relevanceLevels: {
+                                      csaf_base: 'optional',
+                                      csaf_security_incident_response:
+                                        'optional',
+                                      csaf_informational_advisory: 'excluded',
+                                      csaf_security_advisory: 'optional',
+                                      csaf_vex: 'optional',
+                                    },
+                                    i18n: {
+                                      title:
+                                        'CvssV2CollateralDamagePotentialTitle',
+                                      description:
+                                        'CvssV2CollateralDamagePotentialDescription',
+                                    },
+                                  },
+                                  enum: [
+                                    'NONE',
+                                    'LOW',
+                                    'LOW_MEDIUM',
+                                    'MEDIUM_HIGH',
+                                    'HIGH',
+                                    'NOT_DEFINED',
+                                  ],
+                                  metaInfo: {},
+                                  type: 'STRING',
+                                },
+                                {
+                                  key: 'targetDistribution',
+                                  fullName: [
+                                    'vulnerabilities',
+                                    'scores',
+                                    'cvss_v2',
+                                    'targetDistribution',
+                                  ],
+                                  metaData: {
+                                    uiType: 'STRING_ENUM',
+                                    userDocumentation: {
+                                      specification:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
+                                      usage:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
+                                    },
+                                    relevanceLevels: {
+                                      csaf_base: 'optional',
+                                      csaf_security_incident_response:
+                                        'optional',
+                                      csaf_informational_advisory: 'excluded',
+                                      csaf_security_advisory: 'optional',
+                                      csaf_vex: 'optional',
+                                    },
+                                    i18n: {
+                                      title: 'CvssV2TargetDistributionTitle',
+                                      description:
+                                        'CvssV2TargetDistributionDescription',
+                                    },
+                                  },
+                                  enum: [
+                                    'NONE',
+                                    'LOW',
+                                    'MEDIUM',
+                                    'HIGH',
+                                    'NOT_DEFINED',
+                                  ],
+                                  metaInfo: {},
+                                  type: 'STRING',
+                                },
+                                {
+                                  key: 'confidentialityRequirement',
+                                  fullName: [
+                                    'vulnerabilities',
+                                    'scores',
+                                    'cvss_v2',
+                                    'confidentialityRequirement',
+                                  ],
+                                  metaData: {
+                                    uiType: 'STRING_ENUM',
+                                    userDocumentation: {
+                                      specification:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
+                                      usage:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
+                                    },
+                                    relevanceLevels: {
+                                      csaf_base: 'optional',
+                                      csaf_security_incident_response:
+                                        'optional',
+                                      csaf_informational_advisory: 'excluded',
+                                      csaf_security_advisory: 'optional',
+                                      csaf_vex: 'optional',
+                                    },
+                                    i18n: {
+                                      title:
+                                        'CvssV2ConfidentialityRequirementTitle',
+                                      description:
+                                        'CvssV2ConfidentialityRequirementDescription',
+                                    },
+                                  },
+                                  enum: [
+                                    'LOW',
+                                    'MEDIUM',
+                                    'HIGH',
+                                    'NOT_DEFINED',
+                                  ],
+                                  metaInfo: {},
+                                  type: 'STRING',
+                                },
+                                {
+                                  key: 'integrityRequirement',
+                                  fullName: [
+                                    'vulnerabilities',
+                                    'scores',
+                                    'cvss_v2',
+                                    'integrityRequirement',
+                                  ],
+                                  metaData: {
+                                    uiType: 'STRING_ENUM',
+                                    userDocumentation: {
+                                      specification:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
+                                      usage:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
+                                    },
+                                    relevanceLevels: {
+                                      csaf_base: 'optional',
+                                      csaf_security_incident_response:
+                                        'optional',
+                                      csaf_informational_advisory: 'excluded',
+                                      csaf_security_advisory: 'optional',
+                                      csaf_vex: 'optional',
+                                    },
+                                    i18n: {
+                                      title: 'CvssV2IntegrityRequirementTitle',
+                                      description:
+                                        'CvssV2IntegrityRequirementDescription',
+                                    },
+                                  },
+                                  enum: [
+                                    'LOW',
+                                    'MEDIUM',
+                                    'HIGH',
+                                    'NOT_DEFINED',
+                                  ],
+                                  metaInfo: {},
+                                  type: 'STRING',
+                                },
+                                {
+                                  key: 'availabilityRequirement',
+                                  fullName: [
+                                    'vulnerabilities',
+                                    'scores',
+                                    'cvss_v2',
+                                    'availabilityRequirement',
+                                  ],
+                                  metaData: {
+                                    uiType: 'STRING_ENUM',
+                                    userDocumentation: {
+                                      specification:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
+                                      usage:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
+                                    },
+                                    relevanceLevels: {
+                                      csaf_base: 'optional',
+                                      csaf_security_incident_response:
+                                        'optional',
+                                      csaf_informational_advisory: 'excluded',
+                                      csaf_security_advisory: 'optional',
+                                      csaf_vex: 'optional',
+                                    },
+                                    i18n: {
+                                      title:
+                                        'CvssV2AvailabilityRequirementTitle',
+                                      description:
+                                        'CvssV2AvailabilityRequirementDescription',
+                                    },
+                                  },
+                                  enum: [
+                                    'LOW',
+                                    'MEDIUM',
+                                    'HIGH',
+                                    'NOT_DEFINED',
+                                  ],
+                                  metaInfo: {},
+                                  type: 'STRING',
+                                },
+                                {
+                                  key: 'environmentalScore',
+                                  fullName: [
+                                    'vulnerabilities',
+                                    'scores',
+                                    'cvss_v2',
+                                    'environmentalScore',
+                                  ],
+                                  metaData: {
+                                    userDocumentation: {
+                                      specification:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-spec.en.md',
+                                      usage:
+                                        'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v2-usage.en.md',
+                                    },
+                                    relevanceLevels: {
+                                      csaf_base: 'optional',
+                                      csaf_security_incident_response:
+                                        'optional',
+                                      csaf_informational_advisory: 'excluded',
+                                      csaf_security_advisory: 'optional',
+                                      csaf_vex: 'optional',
+                                    },
+                                    i18n: {
+                                      title: 'CvssV2EnvironmentalScoreTitle',
+                                      description:
+                                        'CvssV2EnvironmentalScoreDescription',
+                                    },
+                                  },
+                                  metaInfo: {},
+                                  type: 'NUMBER',
+                                },
+                              ],
+                            },
+                          },
+                          {
+                            key: 'cvss_v3',
+                            fullName: ['vulnerabilities', 'scores', 'cvss_v3'],
+                            metaData: {
+                              title: 'CVSS v3',
+                              description:
+                                'JSON Schema for Common Vulnerability Scoring System version 3.0',
+                              userDocumentation: {
+                                specification:
+                                  'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v3-spec.en.md',
+                                usage:
+                                  'docs/user/vulnerabilities/vulnerability/scores/score/cvss_v3-usage.en.md',
+                              },
+                              relevanceLevels: {
+                                csaf_base: 'best_practice',
+                                csaf_security_incident_response:
+                                  'best_practice',
+                                csaf_informational_advisory: 'excluded',
+                                csaf_security_advisory: 'best_practice',
+                                csaf_vex: 'best_practice',
+                              },
+                              i18n: {
+                                title: 'CvssV3Title',
+                                description: 'CvssV3Description',
+                              },
+                              uiType: 'OBJECT_CVSS_3',
+                            },
+                            type: 'OBJECT',
+                            metaInfo: { propertyList: [] },
+                          },
+                          {
+                            key: 'products',
+                            fullName: ['vulnerabilities', 'scores', 'products'],
+                            title: 'List of product_ids',
+                            description:
+                              'Specifies a list of product_ids to give context to the parent item.',
+                            metaData: {
+                              userDocumentation: {
+                                specification:
+                                  'docs/user/vulnerabilities/vulnerability/scores/score/products-spec.en.md',
+                                usage:
+                                  'docs/user/vulnerabilities/vulnerability/scores/score/products-usage.en.md',
+                              },
+                              relevanceLevels: {
+                                csaf_base: 'mandatory',
+                                csaf_security_incident_response: 'mandatory',
+                                csaf_informational_advisory: 'mandatory',
+                                csaf_security_advisory: 'mandatory',
+                                csaf_vex: 'mandatory',
+                              },
+                              i18n: {
+                                title: 'ProductsTitle',
+                                description: 'ProductsDescription',
+                              },
+                            },
+                            type: 'ARRAY',
+                            metaInfo: {
+                              arrayType: {
+                                key: '',
+                                fullName: [
+                                  'vulnerabilities',
+                                  'scores',
+                                  'products',
+                                ],
+                                title: 'Reference token for product instance',
+                                description:
+                                  'Token required to identify a full_product_name so that it can be referred to from other parts in the document. There is no predefined or required format for the product_id as long as it uniquely identifies a product in the context of the current document.',
+                                metaData: {
+                                  uiType: 'STRING_PRODUCT_ID',
+                                  userDocumentation: {
+                                    specification:
+                                      'docs/user/types/product_id-spec.en.md',
+                                    usage:
+                                      'docs/user/types/product_id-usage.en.md',
+                                  },
+                                  i18n: {
+                                    title: 'ProductIdTitle',
+                                    description: 'ProductIdDescription',
+                                  },
+                                  itemName: {
+                                    itemNameTranslationKey:
+                                      'arrays.productsItemName',
+                                  },
+                                  relevanceLevels: {
+                                    csaf_base: 'mandatory',
+                                    csaf_security_incident_response:
+                                      'mandatory',
+                                    csaf_informational_advisory: 'mandatory',
+                                    csaf_security_advisory: 'mandatory',
+                                    csaf_vex: 'mandatory',
+                                  },
+                                },
+                                minLength: 1,
+                                examples: ['CSAFPID-0004', 'CSAFPID-0008'],
+                                metaInfo: {},
+                                type: 'STRING',
+                              },
+                            },
                           },
                         ],
                       },
