+++ app/scripts/importUiMetaData/csaf_2_0/csaf_json_schema.json
