+++ app/scripts/importUiMetaData/csaf_2_0/types.ts
