+++ app/scripts/importUiMetaData/csaf_2_1/csaf_json_schema.json
